requirements.txt
# Core ML
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
lightgbm>=4.0
xgboost>=2.0
optuna>=3.4

# Geospatial
geopandas>=0.14
rasterio>=1.3
shapely>=2.0
netCDF4>=1.6
xarray>=2023.1

# Visualization
matplotlib>=3.7
cartopy>=0.22
tqdm>=4.65

# Data APIs
requests>=2.31
python-dotenv>=1.0
earthengine-api>=0.1.370
cdsapi>=0.6

# Statistical analysis
scipy>=1.11
